backend/requirements.txt
fastapi
uvicorn
trendspy
feedparser
requests
python-jose[cryptography]
passlib[bcrypt]
google-auth
google-auth-oauthlib
sqlalchemy
